refactor(test): refactor pointer tests
